feat: Add `getMeta` server utility
