+++ src/runtime/server/api/mutation/create.ts
@@ -1,6 +1,6 @@
 import { readMultipartFormData, createError } from 'h3'
 import { defineEventHandler, useRuntimeConfig } from '#imports'
-import { normalizeKey, getKey } from '#s3'
+import { normalizeKey, getKey, getMeta } from '#s3'
 
 export default defineEventHandler(async (event) => {
   const config = useRuntimeConfig()
@@ -10,7 +10,6 @@ export default defineEventHandler(async (event) => {
   const multipartFormData = await readMultipartFormData(event)
 
   const file = multipartFormData?.find(el => el.name === 'file')
-  const meta = multipartFormData?.find(el => el.name === 'meta')
 
   if (file) {
     const normalizedKey = normalizeKey(key)
@@ -18,11 +17,11 @@ export default defineEventHandler(async (event) => {
     verifyType(file.type, config.public.s3.accept)
     verifySize(file.data.length, config.public.s3.maxSizeMb)
 
-    const s3Meta = meta && JSON.parse(meta.data.toString())
+    const s3Meta = await getMeta(event)
 
     await event.context.s3.setItemRaw(normalizedKey, file.data, { s3Meta })
 
-    await event.context.s3.setMeta(normalizedKey, { s3Meta })
+    await event.context.s3.setMeta(normalizedKey, s3Meta)
 
     return { status: 'ok' }
   }
